+++ src/test/java/com/zaychikov/tests/TestData.java
@@ -1,0 +1,4 @@
+package com.zaychikov.tests;
+
+public class TestData {
+}
